# Conflicts:
#	build-all.sh

+++ tpi-admin-service/src/main/java/com/tpi/admin/backup/repositories/NotificacionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificacionRepository extends JpaRepository<Notificacion, Long> {
+public interface NotificacionRepository extends JpaRepository<Notificacion, Integer> {
 }
